techmo_trybun_pyclient/requirements.txt
grpcio==1.0.1
protobuf==3.1.0
